Rename Modal to CityDetailModal to avoid confusion, and move to the appropriate subfolder
