+++ src/components/pages/Comparison/ComparisonCard.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import ModalComponent from "../../common/Modal";
+import CityDetailModal from "../CityDetail/CityDetailModal";
 import { Divider, Skeleton } from "antd";
 //icons
 import weather from "../../../styles/icons/weather-48.png";
@@ -119,7 +119,7 @@ export default class ComparisonCard extends Component {
             </div>
           </div>
         </div>
-        <ModalComponent
+        <CityDetailModal
           visible={visible}
           toggleModal={toggleModal}
           city={this.props.city}
